gac/requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.8.7
django-allauth==0.24.1
oauthlib==1.0.3
python-decouple==3.0
python3-openid==3.0.9
requests==2.8.1
requests-oauthlib==0.5.0
wheel==0.24.0
